+++ airbyte-integrations/connectors/source-mysql/src/main/kotlin/io/airbyte/integrations/source/mysql/MySqlSourceOperations.kt
@@ -102,7 +102,7 @@ class MySqlSourceOperations :
             )
         recordData[MySqlSourceCdcMetaFields.CDC_LOG_POS.id] =
             FieldValueEncoder(
-                0.toDouble(),
+                0,
                 MySqlSourceCdcMetaFields.CDC_LOG_POS.type.jsonEncoder as JsonEncoder<Any>
             )
         if (globalStateValue == null) {
@@ -118,7 +118,7 @@ class MySqlSourceOperations :
             )
         recordData[MySqlSourceCdcMetaFields.CDC_LOG_POS.id] =
             FieldValueEncoder(
-                position.position.toDouble(),
+                position.position,
                 MySqlSourceCdcMetaFields.CDC_LOG_POS.type.jsonEncoder as JsonEncoder<Any>
             )
     }
